[servicelist.py] Bug fix in reloadServicesLists()
[servicelist.py] Bug fix in reloadServicesLists()

Recreate bug as follows:

1) Retrieve lamedb from STB over FTP.
2) Manually edit a channel name.
3) Send lamedb to STB over FTP.
4) Call http://<STB-IP>/web/servicelistreload?mode=0
5) Check channel name.

In the case of the current code the service name does not change.
This makes it impossible to correctly reload service lists.
Only way to get clean reload with current code is to put the STB to sleep (init 4).

The code in this commit cleans enigma2 before the reload so the new service list will not be contaminated by the current one loaded in the RAM.